Implementation/requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
Flask==3.0.3
Flask_Cors==4.0.1
google_api_python_client==2.142.0
google_auth_oauthlib==1.0.0
langdetect==1.0.9
matplotlib==3.7.5
numpy==1.24.3
pandas==2.0.3
protobuf==5.28.2
python_whois==0.9.4
Requests==2.32.3
scikit_learn==1.3.2
tldextract==5.1.2
torch==2.4.0
transformers==4.44.0
seaborn== 0.13.2
tqdm==4.66.5
sqlite3==2.6.0
python==3.8.19